Add catchAsync function

+++ src/api/services/userServices.js
@@ -1,4 +1,5 @@
 const UserRepo = require('../repositories/userRepository');
+const AppError = require('../../utils/appError');
 const pagination = require('../../utils/pagination');
 
 exports.getUsers = async ({ page, amount }) => {
@@ -9,7 +10,15 @@ exports.getUsers = async ({ page, amount }) => {
   return { users };
 };
 
-exports.deleteUser = async (userId) => {
+exports.deleteSelf = async ({ userId, authUserId }) => {
+  if (!(await this.getUserById(userId))) {
+    throw new AppError("Specified user doesn't exist", 400);
+  }
+
+  if (userId !== authUserId) {
+    throw new AppError('You have no permission to delete this user', 403);
+  }
+
   return await UserRepo.deleteById(userId);
 };
